+++ inspectit-gepard-agent/build.gradle
@@ -120,8 +120,8 @@ dependencies {
     testImplementation("com.squareup.okhttp3:okhttp:4.12.0")
     testImplementation("io.opentelemetry.proto:opentelemetry-proto:1.3.2-alpha")
     testImplementation("com.google.protobuf:protobuf-java-util:3.25.4")
-    // Reading Files from Resources easier
-    testImplementation 'commons-io:commons-io:2.16.1'
+    // Reading files from resources easier
+    testImplementation("commons-io:commons-io:2.16.1")
 
     // Otel Java instrumentation that we use and extend during integration tests
     otel("io.opentelemetry.javaagent:opentelemetry-javaagent:${versions.opentelemetryJavaagent}")
