+++ inspectit-gepard-agent/src/test/java/rocks/inspectit/gepard/agent/integrationtest/spring/SpringTestBase.java
@@ -2,6 +2,8 @@
 package rocks.inspectit.gepard.agent.integrationtest.spring;
 
 import java.time.Duration;
+import okhttp3.Call;
+import okhttp3.Request;
 import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.containers.wait.strategy.WaitStrategy;
 import rocks.inspectit.gepard.agent.integrationtest.IntegrationTestBase;
@@ -20,5 +22,14 @@ public abstract class SpringTestBase extends IntegrationTestBase {
   protected WaitStrategy getTargetWaitStrategy() {
     return Wait.forLogMessage(".*Started SpringbootApplication in.*", 1)
         .withStartupTimeout(Duration.ofMinutes(1));
+  }
+
+  protected void sendRequestToTarget(String path) throws Exception {
+    String url = String.format("http://localhost:%d%s", target.getMappedPort(8080), path);
+    Call call = client.newCall(new Request.Builder().url(url).get().build());
+    call.execute();
+
+    // wait for logs
+    Thread.sleep(1000);
   }
 }
